Move worklist-related stuff to /worklist folder
