+++ src/containers/WelcomePage/components/worklist/WorkList.tsx
@@ -10,9 +10,9 @@ import { useTranslation } from 'react-i18next';
 import { useCallback, useState } from 'react';
 import { SingleValue } from '@ndla/select';
 import { TabsV2 } from '@ndla/tabs';
-import { useSearch } from '../../../modules/search/searchQueries';
+import { useSearch } from '../../../../modules/search/searchQueries';
 import WorkListTabContent from './WorkListTabContent';
-import { useSearchConcepts } from '../../../modules/concept/conceptQueries';
+import { useSearchConcepts } from '../../../../modules/concept/conceptQueries';
 import ConceptListTabContent from './ConceptListTabContent';
 
 interface Props {
@@ -29,6 +29,9 @@ const WorkList = ({ ndlaId }: Props) => {
 
   const [sortOptionConcepts, setSortOptionConcepts] = useState('-title');
   const [errorConceptList, setErrorConceptList] = useState<string>();
+  const [filterConceptSubjects, setFilterConceptSubjects] = useState<SingleValue | undefined>(
+    undefined,
+  );
 
   const updateSortOptionConcepts = useCallback((v: string) => setSortOptionConcepts(v), []);
 
@@ -86,6 +89,8 @@ const WorkList = ({ ndlaId }: Props) => {
               isLoading={conceptsLoading}
               error={errorConceptList}
               sortOption={sortOptionConcepts}
+              filterSubject={filterConceptSubjects}
+              setFilterSubject={setFilterConceptSubjects}
             />
           ),
         },
